Functionalize MergeRow.
